shush

+++ cellprofiler/utilities/hdf5_dict.py
@@ -535,10 +535,9 @@ class HDF5Dict(object):
                             hdf5_type = new_dtype
                     else:
                         break
-            hdf5_type_is_int = (
-                np.issubdtype(hdf5_type, int) or
-                (isinstance(hdf5_type, np.dtype) and hdf5_type.kind == 'u'))
-            hdf5_type_is_float = np.issubdtype(hdf5_type, float)
+
+            hdf5_type_is_int = (np.issubdtype(hdf5_type, np.int64) or (isinstance(hdf5_type, np.dtype) and hdf5_type.kind == 'u'))
+            hdf5_type_is_float = np.issubdtype(hdf5_type, np.float64)
             hdf5_type_is_string = not (hdf5_type_is_int or hdf5_type_is_float)
         with self.lock:
             if not self.has_feature(object_name, feature_name):
